segregating code logic
